Missing glyphs at map startup
If a bunch of threads are creating strings at the same time (such as in the style sheet parser), one may create a new dynamic texture, but then have another one that makes use of it outrun the first one.  That results in a missing dynamic texture.
Solution is to just skip right to the scene, so the texture is created next frame no matter what.